refactor: move test logging utils
