+++ packages/scripts/src/commands/app-data/convert/utils/app-data-override.utils.spec.ts
@@ -1,5 +1,5 @@
 import { FlowTypes } from "data-models";
-import { useMockWarningLogger } from "../../../../utils/logging.utils";
+import { useMockWarningLogger } from "../../../../../test/helpers/utils";
 import { assignFlowOverrides } from "./app-data-override.utils";
 
 const TEST_INPUTS: FlowTypes.FlowTypeWithData[] = [
