+++ packages/scripts/src/utils/logging/console-logger.ts
@@ -69,22 +69,3 @@ interface ILogOptions {
   logOnly?: boolean; // specify to log and continue processing instead of exiting
 }
 const defaultLog: ILogOptions = { msg1: "", msg2: "", logOnly: false };
-
-/*************************************************************
- * Test utilties
- *************************************************************/
-
-/** Mock function that will replace default `Logger` function to instead just record any invocations */
-export const mockErrorLogger = jasmine.createSpy("mockErrorLogger", Logger.error);
-export function useMockErrorLogger() {
-  spyOn(Logger, "error").and.callFake(mockErrorLogger);
-  return mockErrorLogger;
-}
-
-/** Mock function that will replace default `Logger` function to instead just record any invocations */
-export const mockWarningLogger = jasmine.createSpy("mockWarningLogger", Logger.warning);
-
-export function useMockWarningLogger() {
-  spyOn(Logger, "warning").and.callFake(mockWarningLogger);
-  return mockWarningLogger;
-}
